+++ lib/src/deprecated/deprecated.dart
@@ -1,2 +1,4 @@
-export '300.dart';
-export '302.dart';
+export 'deprecated_300.dart';
+export 'deprecated_302.dart';
+export 'deprecated_303.dart';
+export 'deprecated_310.dart';