+++ lib/src/deprecated/deprecated_310.dart
@@ -1,0 +1,10 @@
+// Project imports:
+import 'package:zego_uikit_prebuilt_live_streaming/zego_uikit_prebuilt_live_streaming.dart';
+
+const deprecatedTipsV310 = ', '
+    'deprecated since 3.1.0, '
+    'will be removed after 3.5.0'
+    'Migrate Guide:https://pub.dev/documentation/zego_uikit_prebuilt_live_streaming/latest/topics/Migration:%20v3.0-topic.html#310';
+
+@Deprecated('Use ZegoLiveStreamingControllerPKImpl instead$deprecatedTipsV310')
+typedef ZegoLiveStreamingPKController = ZegoLiveStreamingControllerPKImpl;