+++ lib/src/pk/layout/layout.dart
@@ -6,10 +6,10 @@ import 'package:flutter/services.dart';
 import 'package:zego_uikit_prebuilt_live_streaming/src/pk/layout/prefer_gird_layout.dart';
 
 /// width of mixer canvas
-const zegoLiveStreamingPK2MixerCanvasWidth = 972.0;
+const zegoLiveStreamingPKMixerCanvasWidth = 972.0;
 
 /// height of mixer canvas
-const zegoLiveStreamingPK2MixerCanvasHeight = 864.0;
+const zegoLiveStreamingPKMixerCanvasHeight = 864.0;
 
 /// Inheritance of the hybrid layout parent class allows you to return your
 /// custom coordinates and modify the layout of the mixed stream.
@@ -19,8 +19,8 @@ abstract class ZegoLiveStreamingPKMixerLayout {
   /// default is zegoPK2MixerCanvasWidth x zegoPK2MixerCanvasHeight
   Size getResolution() {
     return const Size(
-      zegoLiveStreamingPK2MixerCanvasWidth,
-      zegoLiveStreamingPK2MixerCanvasHeight,
+      zegoLiveStreamingPKMixerCanvasWidth,
+      zegoLiveStreamingPKMixerCanvasHeight,
     );
   }